refactor(BeedataApiClient): improved interface
* http session in client
* _request is now a clojure
* clean connections when client is deleted
* download_report returns only one report
