+++ Pipfile
@@ -40,3 +40,4 @@ pyyaml = "*"
 sentry-sdk = "*"
 aiohttp = "*"
 hiredis = "*"
+ipdb = "*"
