+++ infoenergia_api/contrib/__init__.py
@@ -1,4 +1,3 @@
-from .beedata_api import ApiException, BeedataApiClient
 from .cch import BaseCch as Cch
 from .cch import TgCchF1, TgCchF5d, TgCchP1, TgCchVal
 from .contracts import Contract
